+++ src/components/Hero/components/Mobile/index.js
@@ -7,7 +7,7 @@ const HeroMobile = () => {
       <div className="col col-lg-9 col-md-7 slide-caption mobile-only">
         <div className="slide-title mobile-only">
           <h5 className="header-first-half">Indiferent de călătorie, mașina ta a fost mereu alături de tine.</h5>
-          <h5>Respect-o și pregătește-o pentru noi aventuri la <span className="hashtag-alpha-auto">#AlphaAuto</span></h5>
+          <h5 className="header-second-half">Respect-o și pregătește-o pentru noi aventuri la <span className="hashtag-alpha-auto">#AlphaAuto</span></h5>
         </div>
         <div className="mobile-buttons-container">
           <div className="mobile-only localisation-button">